pyproject.toml
[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"

[project]
name = "django-channel-tasks"
version = "0.9.3.dev0"
description = "Running background tasks through websocket, using channels-redis, from Django Admin or REST API."
readme = "README.md"
requires-python = ">=3.11"
license = {file = "LICENSE"}
authors = [
  {name = "Daniel Farré Manzorro", email = "d.farre.m@gmail.com" }
]
maintainers = [
  {name = "Daniel Farré Manzorro", email = "d.farre.m@gmail.com" }
]
classifiers = [
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Intended Audience :: Developers",
    "Development Status :: 4 - Beta",
    "Framework :: Django :: 5.1",
]
dependencies = [
    "Django",
    "django-filter",
    "django-extensions",
    "django-request-logging",
    "adrf",
    "channels",
    "channels-redis",
    "tzdata",
    "psycopg[pool]",
    "websocket-client",
    "django-sass-compiler",
]

[project.optional-dependencies]
dev = [
    "ipdb",
    "ipython",
]
mypy = [
    "mypy",
    "django-stubs",
    "djangorestframework-stubs[compatible-mypy]",
    "types-beautifulsoup4",
    "types-setuptools",
    "flake8",
]
docs = [
    "sphinx",
    "pydeps",
    "pydot",
    "pygraphviz",
]
test = [
    "pytest",
    "pytest-cov",
    "pytest-django",
    "pytest-asyncio",
    "pytest-timeout",
    "bdd-coder==2.2.3.dev3",
    "beautifulsoup4",
    "requests",
]

[project.scripts]
channel-tasks-admin = "django_tasks.entrypoint:manage_channel_tasks"

[project.urls]
"Homepage" = "https://github.com/dfarre/django-channel-tasks"
"Source" = "https://github.com/dfarre/django-channel-tasks"
"Issue Tracker" = "https://github.com/dfarre/django-channel-tasks/issues"

[tool.setuptools.package-data]
"django_tasks" = [
    "templates/task_alerts.html",
    "templates/admin/base.html",
    "static/css/bootstrap-isolated.min.css",
]

[tool.setuptools.data-files]
"./channel-tasks-docker" = [
    "docker-compose.yml",
    "Dockerfile",
    "setup-channel-tasks-docker.sh",
    "channel-tasks-unit.template.json",
]
"./channel-tasks-unit" = [
    "services/nginx-unit/application.template.json",
    "services/nginx-unit/listener.template.json",
    "services/nginx-unit/routes.template.json",
]
"channel-tasks-services" = [
    "services/channel-tasks-docker.template.service",
    "services/channel-tasks.template.service",
]
"./bin" = [
    "services/restart-channel-tasks-docker.sh",
    "services/setup-channel-tasks.sh",
]
